Remove unused @govuk-react/storybook-components
