+++ components/radio/src/stories.tsx
@@ -1,10 +1,12 @@
 import type { FieldInputProps } from 'react-final-form';
 
 import React from 'react';
-import { Field } from 'react-final-form';
+import { Form, Field } from 'react-final-form';
 
+import { action } from '@storybook/addon-actions';
+
+import Button from '@govuk-react/button';
 import { withKnobs } from '@storybook/addon-knobs';
-import { FinalFormWrapper } from '@govuk-react/storybook-components';
 
 import MultiChoice from '@govuk-react/multi-choice';
 import { Radio } from '.';
@@ -125,6 +127,34 @@ export const RadioWithHintText: React.FC = () => (
     </Radio>
   </div>
 );
+
+
+const FinalFormWrapper: React.FC = ({ children }: { children: React.ReactNode }) => (
+  <Form
+    onSubmit={action('submit')}
+    render={({ handleSubmit, form: { reset }, submitting, pristine, values }) => (
+      <form onSubmit={handleSubmit}>
+        <div>{children}</div>
+        <div>
+          <Button type="submit" disabled={submitting}>
+            Submit
+          </Button>
+        </div>
+        <div>
+          <Button onClick={reset} disabled={submitting || pristine}>
+            Reset
+          </Button>
+        </div>
+        <div>
+          <hr />
+          <pre>{JSON.stringify(values, null, 2)}</pre>
+        </div>
+      </form>
+    )}
+  />
+);
+
+
 
 export const UsageWithFinalReduxFormMultiCheckboxValidation: React.FC = () => (
   <FinalFormWrapper>
