+++ jest.config.js
@@ -11,7 +11,6 @@ module.exports = {
     '!<rootDir>/{components,packages}/*/src/**/{stories,example,fixtures,test}.[jt]s?(x)',
     '!<rootDir>/packages/storybook/**',
     '!<rootDir>/packages/api-docs/**',
-    '!<rootDir>/packages/storybook-components/**',
     '!<rootDir>/packages/example-application/**',
   ],
   coverageDirectory: './coverage/',
