--- packages/storybook-components/.eslintrc.js
@@ -1,5 +1,0 @@
-module.exports = {
-  "env": {
-    "browser": true
-  }
-}
