--- packages/storybook-components/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@govuk-react/storybook-components",
-  "version": "0.9.1",
-  "dependencies": {
-    "@govuk-react/button": "^0.9.1",
-    "@storybook/addon-actions": "6.3.5",
-    "@storybook/addon-docs": "6.3.5",
-    "@storybook/addons": "6.3.7",
-    "@storybook/react": "6.3.6",
-    "react-final-form": "^6.5.3"
-  },
-  "peerDependencies": {
-    "react": ">=16.2.0"
-  },
-  "main": "dist/govuk-react-storybook-components.cjs.js",
-  "module": "dist/govuk-react-storybook-components.esm.js",
-  "source": "src/index.tsx",
-  "author": "Alasdair McLeay",
-  "license": "MIT",
-  "homepage": "https://github.com/govuk-react/govuk-react#readme",
-  "description": "govuk-react: A port of the govuk-frontend components for React using styled-components.",
-  "private": false,
-  "publishConfig": {
-    "access": "public"
-  }
-}
