--- packages/storybook-components/src/index.tsx
@@ -1,3 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-export { default as FinalFormWrapper } from './final-form-wrapper';
-export { default as withDocsCustom } from './with-docs-custom';
