--- packages/storybook-components/src/with-docs-custom.tsx
@@ -1,8 +1,0 @@
-// TODO: remove all uses
-
-const withDocsCustom =
-  () =>
-  (storyFn: () => void): void =>
-    storyFn();
-
-export default withDocsCustom;
