+++ scripts/createComponent.js
@@ -83,8 +83,6 @@ const storiesScript = () => {
   const filename = 'stories.js';
   const contents = `import React from 'react';
 import { storiesOf } from '@storybook/react';
-// TODO: remove comments for documentation once docs have been generated
-// import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import ${componentName} from '.';
 
@@ -94,12 +92,9 @@ const stories = storiesOf('${componentName}', module);
 
 stories.add(
   'Component default',
-  // withDocsCustom(
-  //   ReadMe,
     () => (
       <${componentName}>${componentName} example</${componentName}>
     ),
-  // ),
 );
 `;
   writeFile(filename, contents);
